bug de edicion de producto

+++ database/migrations/2023_02_23_193216_create_products_table.php
@@ -24,7 +24,7 @@ class CreateProductsTable extends Migration
 
             $table->text('description');
 
-            $table->float('price');
+            $table->decimal('price',20);
             
             $table->unsignedBigInteger('subcategory_id');
             $table->foreign('subcategory_id')->references('id')->on('subcategories')->onDelete('cascade');
